Initial MVC implementation: creating router and database class
