+++ .gitignore
@@ -3,3 +3,4 @@ config.py
 __pycache__/
 .vscode/
 key.py
+.env
