+++ backend/routers/__init__.py
@@ -1,0 +1,5 @@
+from .attractions import *
+from . user import *
+from .booking import *
+from .orders import *
+
