+++ frontend/CSS/attraction.css
