+++ frontend/assets/icon_delete.png
