+++ frontend/assets/left_arrow_default.png
