+++ frontend/assets/left_arrow_hovered.png
