+++ frontend/assets/right_arrow_default.png
