+++ frontend/assets/right_arrow_hovered.png
