+++ frontend/static/attraction.html
@@ -9,8 +9,8 @@
   <link
     href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100..900&family=Noto+Sans:ital,wght@0,100..900;1,100..900&display=swap"
     rel="stylesheet">
-  <link rel="stylesheet" href="../external/attraction.css">
-  <link rel="stylesheet" href="../external/login.css">
+  <link rel="stylesheet" href="../css/attraction.css">
+  <link rel="stylesheet" href="../css/login.css">
 </head>
 <body>
   <nav>
@@ -24,7 +24,7 @@
   </nav>
 
   <dialog id="login-dialogue">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -40,7 +40,7 @@
   </dialog>
 
   <dialog id="signup-dialog">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -60,8 +60,8 @@
     <div class="showcase">
       <div class="gallery">
         <div class="arrow">
-          <img id="leftArrow" src="../external/left_arrow_hovered.png">
-          <img id="rightArrow" src="../external/right_arrow_hovered.png">
+          <img id="leftArrow" src="../assets/left_arrow_hovered.png">
+          <img id="rightArrow" src="../assets/right_arrow_hovered.png">
         </div>
         <div class="slider">
        <!-- Render the gallery here -->
@@ -113,9 +113,9 @@
   </footer>
 
 
-  <script src="../external/attraction.js" defer></script>
-  <script src="../external/login.js" defer></script>
-  <script src="../external/credential.js" defer></script>
-  <script src="../external/bookingBtn.js" defer></script>
+  <script src="../js/attraction.js" defer></script>
+  <script src="../js/login.js" defer></script>
+  <script src="../js/credential.js" defer></script>
+  <script src="../js/bookingBtn.js" defer></script>
 </body>
 </html>
