+++ frontend/static/booking.html
@@ -5,8 +5,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Booking</title>
-  <link rel="stylesheet" href="../external/booking.css">
-  <link rel="stylesheet" href="../external/login.css">
+  <link rel="stylesheet" href="../css/booking.css">
+  <link rel="stylesheet" href="../css/login.css">
 </head>
 <style>
   .tpfield {
@@ -45,7 +45,7 @@
   </nav>
 
   <dialog id="login-dialogue">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -61,7 +61,7 @@
   </dialog>
 
   <dialog id="signup-dialog">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -83,7 +83,7 @@
       <img src="" id="mainImage">
       <div class="attractionInfo">
         <h4>台北一日遊：<span id="attractionName">平安鐘</span></h4>
-        <img src="../external/icon/icon/icon_delete.png" id="deleteBtn">
+        <img src="../assets/icon_delete.png" id="deleteBtn">
         <p id=><span class="title">日期：</span><span id="bookingDate">2025年4月15日</span></p>
         <p><span class="title">時間：</span><span id="bookingTime">早上9點到下午4點</span></p>
         <p><span class="title">費用：</span>新台幣<span id="tourCost">2000</span>元</p>
@@ -129,11 +129,11 @@
     <p>COPYRIGHT © 2025 台北一日遊</p>
   </footer>
 </body>
-<script src="../external/login.js"></script>
-<script src="../external/credential.js"></script>
-<script src="../external/bookingBtn.js"></script>
-<script src="../external/booking.js"></script>
+<script src="../js/login.js"></script>
+<script src="../js/credential.js"></script>
+<script src="../js/bookingBtn.js"></script>
+<script src="../js/booking.js"></script>
 <script src="https://js.tappaysdk.com/sdk/tpdirect/v5.19.2" defer></script>
-<script src="../external/tappay.js" defer></script>
+<script src="../js/tappay.js" defer></script>
 
 </html>
