+++ frontend/static/index.html
@@ -10,8 +10,8 @@
   <link
     href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100..900&family=Noto+Sans:ital,wght@0,100..900;1,100..900&display=swap"
     rel="stylesheet">
-  <link rel="stylesheet" href="../external/index.css">
-  <link rel="stylesheet" href="../external/login.css">
+  <link rel="stylesheet" href="../css/index.css">
+  <link rel="stylesheet" href="../css/login.css">
 </head>
 
 <body>
@@ -26,7 +26,7 @@
   </nav>
 
   <dialog id="login-dialogue">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -42,7 +42,7 @@
   </dialog>
 
   <dialog id="signup-dialog">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -67,7 +67,7 @@
           <input type="text" name="keyword" id="searchBox" placeholder="輸入景點名稱查詢">
         </form>
           <div id="searchIcon">
-            <img src="../external/icon_search.png" alt="search icon" id="searchImage">
+            <img src="../assets/icon_search.png" alt="search icon" id="searchImage">
           </div>
         
       </div>
@@ -76,11 +76,11 @@
 
   <main>
     <div id="scroll">
-      <img src="../external/left arrow.png" id="leftClick">
+      <img src="../assets/left arrow.png" id="leftClick">
       <div class="list-bar">
         <!-- Append the MRT stations here -->
       </div>
-      <img src="../external/right arrow.png" id="rightClick">
+      <img src="../assets/right arrow.png" id="rightClick">
     </div>
     
     <div class="display">
@@ -94,10 +94,10 @@
     <p>COPYRIGHT © 2025 台北一日遊</p>
   </footer>
   <div id="sentinel"></div>
-  <script src="../external/index.js"></script>
-  <script src="../external/login.js"></script>
-  <script src="../external/credential.js"></script>
-  <script src="../external/bookingBtn.js"></script>
+  <script src="../js/index.js"></script>
+  <script src="../js/login.js"></script>
+  <script src="../js/credential.js"></script>
+  <script src="../js/bookingBtn.js"></script>
 </body>
 
 </html>
