+++ frontend/static/thankyou.html
@@ -9,9 +9,9 @@
   <link
     href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@100..900&family=Noto+Sans:ital,wght@0,100..900;1,100..900&display=swap"
     rel="stylesheet">
-  <link rel="stylesheet" href="../external/thankyou.css">
-  <link rel="stylesheet" href="../external/login.css">
-  
+  <link rel="stylesheet" href="../css/thankyou.css">
+  <link rel="stylesheet" href="../css/login.css">
+    
 </head>
 <body>
   <nav>
@@ -25,7 +25,7 @@
   </nav>
 
   <dialog id="login-dialogue">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -41,7 +41,7 @@
   </dialog>
 
   <dialog id="signup-dialog">
-    <img src="../external/icon/icon_close.png">
+    <img src="../assets/icon_close.png">
     <div>
       <div id="decorator-bar"></div>
       <div class="main-box">
@@ -82,8 +82,8 @@
 <footer class="highfooter">
   <p>COPYRIGHT © 2025 台北一日遊</p>
 </footer>
-<script src="../external/thankyou.js"></script>
-<script src="../external/login.js"></script>
-<script src="../external/credential.js"></script>
-<script src="../external/bookingBtn.js"></script>
+<script src="../js/thankyou.js"></script>
+<script src="../js/login.js"></script>
+<script src="../js/credential.js"></script>
+<script src="../js/bookingBtn.js"></script>
 </html>
