+++ requirements.txt
@@ -35,6 +35,7 @@ matplotlib-inline==0.1.7
 mdurl==0.1.2
 mysql-connector-python==9.2.0
 nest-asyncio==1.6.0
+orjson==3.10.16
 packaging==24.2
 parso==0.8.4
 pexpect==4.9.0
@@ -46,6 +47,8 @@ pure-eval==0.2.3
 pycparser==2.22
 pydantic==2.10.6
 pydantic-core==2.27.2
+pydantic-extra-types==2.10.3
+pydantic-settings==2.9.1
 pygments==2.19.1
 pyjwt==2.10.1
 python-dateutil==2.9.0.post0
@@ -58,12 +61,16 @@ rich-toolkit==0.13.2
 shellingham==1.5.4
 six==1.17.0
 sniffio==1.3.1
+sqlalchemy==2.0.40
+sqlmodel==0.0.24
 stack-data==0.6.3
 starlette==0.46.1
 tornado==6.4.2
 traitlets==5.14.3
 typer==0.15.2
 typing-extensions==4.12.2
+typing-inspection==0.4.0
+ujson==5.10.0
 uvicorn==0.34.0
 uvloop==0.21.0
 watchfiles==1.0.4
