+++ sample.env
@@ -1,0 +1,14 @@
+#Create a .env file that includes the following:
+
+# For MySQL database
+config = '{"user": "db_credential", "password": "password_to_db", "host": "local_host", "database": "name_of_the_db"}'
+
+#For password hashing
+secret_key = "deu1098129jdu1092u3ioi109283" #Arbitrary secret key for your hashing algorithm
+algorithm = "HS256" #Or your preferred hashing algorithm
+
+#TapPay Related API Keys
+x-api-key = "partner_fowejfoiejfojweoifejewf"  
+partner_key = "partner_djiodjowqjdoijwqodjowq"
+merchant_id = "tppf_djiqdojojoi_GP_POS_2"
+tappay_URL = "https://sandbox.tappaysdk.com/tpc/payment/pay-by-prime"
